refactor(readme): update path to cmdline tools
From version 4 of command line tools the path to binaries was updated to `/cmdline-tools/latest/bin` with the previous being deprecated. The issue is the deprecated path still exists but does not work with Java v11+ which makes this getting started doc broken and also gives errors not making clear the issue.